// src/types/zipCodes.ts
// Generated by https://quicktype.io
export interface ZipCodeData {
    avatars?:           string[];
    buyer_cost?:        number;
    buyer_max_budget?:  number;
    buyer_share?:       number;
    city?:              string;
    county?:            string;
    density?:           string;
    home_value?:        number;
    id?:                number;
    nearby_zipcodes?:   nearbyZipCodes[];
    lat?:               string;
    lng?:               string;
    population?:        number;
    seller_cost?:       number;
    seller_max_budget?: number;
    seller_share?:      number;
    state_long?:        string;
    state_short?:       string;
    zip?:               string;
}

export interface nearbyZipCodes {
    dist?: number;
    zip?: string;
}